added a autofill combobox for player search
