+++ client/package.json
@@ -33,7 +33,7 @@
     "css-loader": "^5.0.0",
     "file-loader": "^6.2.0",
     "foundation-sites": "^6.6.3",
-    "html-webpack-plugin": "^4.5.0",
+    "html-webpack-plugin": "^5.1.0",
     "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.14.1",
     "prop-types": "~15.7",
