+++ server/src/routes/api/v1/tablesRouter.js
@@ -101,7 +101,6 @@ tablesRouter.patch("/:tableId", async (req, res) => {
     if (error instanceof ValidationError) {
       return res.status(422).json({ errors: error.data })
     }
-    console.log(error)
     return res.status(500).json({ errors: error })
   }
 })
