+++ server/src/services/findStatAndRelateToTable.js
@@ -1,9 +1,7 @@
 import { Stat } from "../models/index.js"
 
 const findStatAndRelateToTable = async (stat, table) => {
-  debugger
   let currentStat = await Stat.query().where({ id: stat.id })
-  debugger
   await table.$relatedQuery("stats").relate(currentStat)
 }
